+++ src/execution/__tests__/defer-test.ts
@@ -22,7 +22,7 @@ import { execute, experimentalExecuteIncrementally } from '../entrypoints.js';
 import type {
   InitialIncrementalExecutionResult,
   SubsequentIncrementalExecutionResult,
-} from '../execute.js';
+} from '../Executor.js';
 
 const friendType = new GraphQLObjectType({
   fields: {
